requirements.txt
annotated-types==0.7.0
attrs==24.2.0
boto3==1.35.52
botocore==1.35.52
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
outcome==1.3.0.post0
packaging==24.1
playwright==1.48.0
playwright-stealth==1.0.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
selenium==4.25.0
selenium-stealth==1.0.6
sentry-sdk==2.17.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tiktok-captcha-solver==0.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
